Timer fixes

+++ src/game/chars/CCharMemory.cpp
@@ -417,7 +417,7 @@ bool CChar::Memory_Fight_OnTick( CItemMemory * pMemory )
 	// Attacker.Elapsed = -1 means no combat end for this attacker.
 	// g_Cfg.m_iAttackerTimeout = 0 means attackers doesnt decay. (but cleared when the attacker is killed or the char dies)
 
-	if (GetDist(pTarg) > UO_MAP_VIEW_RADAR || (g_Cfg.m_iAttackerTimeout != 0 && elapsed >= 0 && (unsigned int)elapsed > g_Cfg.m_iAttackerTimeout))
+	if  (GetDist(pTarg) > UO_MAP_VIEW_RADAR || ((g_Cfg.m_iAttackerTimeout != 0) && (elapsed >= 0) && (elapsed > g_Cfg.m_iAttackerTimeout)) )
 	{
 		Memory_Fight_Retreat( pTarg, pMemory );
 	clearit:
